First version.
Available urls:

<xyz>:8080/UserManagement/rest/TestService/tests
Gives json with all tests.

<xyz>:8080/UserManagement/rest/TestService/tests/<123>
Gives json of a certain test.

<xyz>:8080/UserManagement/rest/TestService/tests/<123>/name
Gives name of a certain test (in plain text)

<xyz>:8080/UserManagement/rest/TestService/tests/uniqueIds
Unique id of every test in one json.

<xyz>:8080/UserManagement/rest/TestService/tests/notifications
This is not finished, in the future planned as notification system for
Speechlist frontend, informing about deprecated app version or temporary
server problems.